Move tests to test_ir. Add doc-string
